api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow==10.1.0
python-dotenv==1.0.0
pydantic==2.5.0
openai==1.12.0
requests==2.31.0
httpx>=0.28.0
redis==5.0.1
aioredis==2.0.1 